+++ LuckyoneUI/Profiles/ElvUI.lua
@@ -1837,8 +1837,8 @@ function Private:Setup_ElvUI(layout)
 	E.db.unitframe.units.raid1.phaseIndicator.anchorPoint = 'TOP'
 	E.db.unitframe.units.raid1.phaseIndicator.scale = 0.5
 	E.db.unitframe.units.raid1.power.enable = false
+	E.db.unitframe.units.raid1.privateAuras.icon.offset = 2
 	E.db.unitframe.units.raid1.privateAuras.icon.point = 'TOP'
-	E.db.unitframe.units.raid1.privateAuras.parent.offsetX = -3
 	E.db.unitframe.units.raid1.privateAuras.parent.point = 'BOTTOMRIGHT'
 	E.db.unitframe.units.raid1.pvpclassificationindicator.enable = false
 	E.db.unitframe.units.raid1.raidRoleIcons.combatHide = true
@@ -1886,6 +1886,7 @@ function Private:Setup_ElvUI(layout)
 	E.db.unitframe.units.raid3.phaseIndicator.anchorPoint = 'LEFT'
 	E.db.unitframe.units.raid3.phaseIndicator.scale = 0.5
 	E.db.unitframe.units.raid3.privateAuras.enable = false
+	E.db.unitframe.units.raid3.privateAuras.icon.offset = 2
 	E.db.unitframe.units.raid3.privateAuras.parent.offsetX = -20
 	E.db.unitframe.units.raid3.privateAuras.parent.point = 'RIGHT'
 	E.db.unitframe.units.raid3.pvpclassificationindicator.enable = false
@@ -2045,9 +2046,11 @@ function Private:Setup_ElvUI(layout)
 
 		-- Main Private Auras
 		E.db.unitframe.units.raid1.privateAuras.icon.size = 16
-		E.db.unitframe.units.raid1.privateAuras.parent.offsetY = 18
+		E.db.unitframe.units.raid1.privateAuras.parent.offsetX = 5
+		E.db.unitframe.units.raid1.privateAuras.parent.offsetY = 26
 		E.db.unitframe.units.raid2.privateAuras.icon.size = 16
-		E.db.unitframe.units.raid2.privateAuras.parent.offsetY = 18
+		E.db.unitframe.units.raid2.privateAuras.parent.offsetX = 5
+		E.db.unitframe.units.raid2.privateAuras.parent.offsetY = 26
 
 		-- Main misc
 		E.global.datatexts.customPanels.Luckyone_ActionBars_DT.width = 395
@@ -2142,11 +2145,13 @@ function Private:Setup_ElvUI(layout)
 		E.db.unitframe.units.raid3.rdebuffs.yOffset = 11
 		E.db.unitframe.units.raid3.width = 140
 
-		-- Main Private Auras
+		-- Healing Private Auras
 		E.db.unitframe.units.raid1.privateAuras.icon.size = 20
-		E.db.unitframe.units.raid1.privateAuras.parent.offsetY = 22
+		E.db.unitframe.units.raid1.privateAuras.parent.offsetX = 7
+		E.db.unitframe.units.raid1.privateAuras.parent.offsetY = 32
 		E.db.unitframe.units.raid2.privateAuras.icon.size = 20
-		E.db.unitframe.units.raid2.privateAuras.parent.offsetY = 22
+		E.db.unitframe.units.raid2.privateAuras.parent.offsetX = 7
+		E.db.unitframe.units.raid2.privateAuras.parent.offsetY = 32
 
 		-- Healing misc
 		E.global.datatexts.customPanels.Luckyone_ActionBars_DT.width = 704